+++ tests/testthat/testRbuildignore/R/ignore_me.R
@@ -1,0 +1,3 @@
+#' Ignore me, I'm not ready
+#'
+ignore_me <- "ignore me"